introduce registry for interpolators
