+++ experanto/interpolators/__init__.py
@@ -1,0 +1,3 @@
+from .base import Interpolator
+from .screen_interpolator import ScreenInterpolator
+from .sequence_interpolator import SequenceInterpolator, PhaseShiftedSequenceInterpolator
